pyproject.toml
[project]
name = "chunk2milvus-hq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pymilvus>=2.3.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
